server/python/requirements.txt
mss
opencv-python
numpy
pyautogui
pillow
tk-tools